requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.24.2
google-auth==2.40.1
google-genai==1.16.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scapy==2.6.1
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
